IEP-445: Refactored the project name from tests to test
