+++ tests/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>com.espressif.idf.tests</name>
+	<name>com.espressif.idf.test</name>
 	<comment></comment>
 	<projects>
 	</projects>
