+++ tests/pom.xml
@@ -5,7 +5,7 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.espressif.idf</groupId>
-	<artifactId>com.espressif.idf.tests</artifactId>
+	<artifactId>com.espressif.idf.test</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
