+++ tests/src/com/espressif/idf/test/common/configs/DefaultPropertyFetcher.java
@@ -2,7 +2,7 @@
  * Copyright 2021 Espressif Systems (Shanghai) PTE LTD. All rights reserved.
  * Use is subject to license terms.
  *******************************************************************************/
-package com.espressif.idf.tests.common.configs;
+package com.espressif.idf.test.common.configs;
 
 import java.io.IOException;
 import java.io.InputStream;
