+++ tests/src/com/espressif/idf/test/operations/SWTBotTreeOperations.java
@@ -2,7 +2,7 @@
  * Copyright 2021 Espressif Systems (Shanghai) PTE LTD. All rights reserved.
  * Use is subject to license terms.
  *******************************************************************************/
-package com.espressif.idf.tests.operations;
+package com.espressif.idf.test.operations;
 
 import java.util.Arrays;
 import java.util.Optional;
